refactor(rr): Split build plan lowering to two files
